+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 2026.4.0
 
+### Note
+- config に `threadPoolSize` オプションが追加されました。
+  - デフォルトは `1` で、ワーカーごとに指定した数のスレッドが作成されます。
+  - スレッドプールは CPU バウンドな処理をオフロードするために使用されるため、みだりに大きな値を指定しないでください。
+
 ### General
 - Enhance: アバターデコレーションにカテゴリを設定できるように
 
@@ -15,6 +20,7 @@
   (Cherry-picked from https://github.com/MisskeyIO/misskey/pull/1410)
 - Enhance: バックエンドの開発モード時の安定性向上
 - Enhance: バックエンドビルド・テスト時に使用する依存関係の整理（swc/esbuild→Rolldown, Jest→Vitest）
+- Enhance: RSA 署名処理のオフロード
 - Fix: ファイルシステムを用いる処理におけるパスの取り扱いを改善
 - Fix: `/api-doc` にアクセスできない問題を修正
 - Fix: support `alsoKnownAs` from remote actors as either array or unwrapped singleton